+++ serializable.ts
@@ -32,7 +32,7 @@ export declare interface ToJSON {
 
 /** reutrns  a new javascript object with transformations */
 export declare interface FromJSON {
-  fromJSON(json: string | JSONValue | Object): this;
+  fromJSON(json: JSONValue): this;
 }
 
 /** returns the javascript object as a `JSONObject` with transformations */
@@ -40,9 +40,16 @@ export declare interface Serialize {
   tsSerialize(): JSONObject;
 }
 
+/** deep copy `this`, jsonObject is a POJO of the class that overrides the cloned
+ * object, jsonObject keys do not need keyTransforms, and values are raw JS Objects
+ */
+export declare interface Clone {
+  clone(jsonObject: Partial<this>): this;
+}
+
 /** Recursively set default serializer logic for own class definition and parent definitions if none exists */
 function getOrInitializeDefaultSerializerLogicForParents(
-  targetPrototype: any, // This will the the class instance, regardless of how low level it is
+  targetPrototype: Serializable,
 ): SerializePropertyOptionsMap | undefined {
   // Don't create serialization logic for Serializable
   if (targetPrototype === Serializable.prototype) {
@@ -87,21 +94,29 @@ function getOrInitializeDefaultSerializerLogicForParents(
  *       example.tsSerialize()
  */
 export abstract class Serializable {
-  /** allow empty class serialization */
   constructor() {
     getOrInitializeDefaultSerializerLogicForParents(this.constructor.prototype);
   }
-  public tsTransformKey?(key: string): string {
+  public tsTransformKey(key: string): string {
     return key;
   }
   public toJSON(): string {
     return toJSON(this);
   }
-  public fromJSON(json: string | JSONValue | Object): this {
+  public fromJSON(json: JSONValue): this {
     return fromJSON(this, json);
   }
   public tsSerialize(): JSONObject {
     return toPojo(this);
+  }
+  public clone(jsonObject: Partial<this> = {}): this {
+    const copy = Object.getPrototypeOf(this).constructor;
+    return Object.assign(
+      new copy().fromJSON(
+        this.tsSerialize(),
+      ),
+      jsonObject,
+    );
   }
 }
 
@@ -116,7 +131,7 @@ export const SERIALIZABLE_CLASS_MAP: SerializableMap = new Map<
 
 /** Converts to object using mapped keys */
 export function toPojo(
-  context: any,
+  context: Serializable,
 ): JSONObject {
   const serializablePropertyMap = SERIALIZABLE_CLASS_MAP.get(
     context?.constructor?.prototype,
@@ -137,25 +152,17 @@ export function toPojo(
     } of serializablePropertyMap.propertyOptions()
   ) {
     // Assume that key is always a string, a check is done earlier in SerializeProperty
-    const value = context[propertyKey as string];
+    const value = context[propertyKey as keyof Serializable];
 
     // If the value is serializable then use the recursive replacer
     if (
       SERIALIZABLE_CLASS_MAP.get(
-        (value as Serializable)?.constructor?.prototype,
+        value?.constructor?.prototype,
       )
     ) {
       toJSONStrategy = toJSONRecursive;
     }
-
-    if (Array.isArray(value)) {
-      record[serializedKey] = value.map((item: any) => {
-        if (item instanceof Serializable) {
-          return toPojo(item);
-        }
-        return toJSONStrategy(item);
-      });
-    } else if (value !== undefined) {
+    if (value !== undefined) {
       record[serializedKey] = toJSONStrategy(value);
     }
   }
@@ -163,33 +170,29 @@ export function toPojo(
 }
 
 /** Convert to `pojo` with our mapping logic then to string */
-function toJSON<T>(context: T): string {
+function toJSON(context: Serializable): string {
   return JSON.stringify(toPojo(context));
 }
 
 /** Convert from object/string to mapped object on the context */
 function fromJSON<T>(
   context: Serializable,
-  json: string | JSONValue | Object,
+  json: JSONValue,
 ): T {
   const _json = typeof json === "string" ? JSON.parse(json) : json;
   const accumulator: Partial<T> = {};
-
-  for (const [key, value] of Object.entries(_json)) {
+  const map = SERIALIZABLE_CLASS_MAP.get(context?.constructor?.prototype);
+  for (const [key, value] of Object.entries(_json) as [string, JSONValue][]) {
     const {
       propertyKey,
       fromJSONStrategy = fromJSONDefault,
-    } = SERIALIZABLE_CLASS_MAP.get(
-      context?.constructor?.prototype,
-    )?.getBySerializedKey(key) || {};
+    } = map?.getBySerializedKey(key) || {};
 
     if (!propertyKey) {
       continue;
     }
 
-    accumulator[propertyKey as keyof T] = Array.isArray(value)
-      ? value.map((v) => fromJSONStrategy(v))
-      : fromJSONStrategy(value as JSONValue);
+    accumulator[propertyKey as keyof T] = fromJSONStrategy(value);
   }
 
   return Object.assign(