+++ serialize_property_test.ts
@@ -7,7 +7,7 @@ import {
   fail,
   test,
 } from "./test_deps.ts";
-import { Serializable } from "./serializable.ts";
+import { JSONValue, Serializable } from "./serializable.ts";
 import { SerializeProperty } from "./serialize_property.ts";
 import {
   ERROR_DUPLICATE_SERIALIZE_KEY,
@@ -58,7 +58,7 @@ test({
   fn() {
     try {
       const TEST = Symbol("test");
-      class Test extends Serializable {
+      class _Test extends Serializable {
         @SerializeProperty()
         [TEST] = "toJSON";
       }
@@ -195,7 +195,7 @@ test({
   fn() {
     class Test extends Serializable {
       @SerializeProperty()
-      array!: any[];
+      array!: unknown[];
     }
     const testObj = new Test().fromJSON(
       `{"array":["worked",0,{"subObj":["cool"]}]}`,
@@ -216,7 +216,8 @@ test({
     }
     class Test extends Serializable {
       @SerializeProperty({
-        fromJSONStrategy: (v) => new OtherClass().fromJSON(v),
+        fromJSONStrategy: (arr) =>
+          (arr as JSONValue[]).map((v) => new OtherClass().fromJSON(v)),
       })
       array!: OtherClass[];
     }
@@ -226,6 +227,19 @@ test({
     assertEquals(testObj.array.length, 5);
     assert(testObj.array[0] instanceof OtherClass);
     assertEquals(testObj.array[4].id, 5);
+  },
+});
+
+test({
+  name: "custom toJSONStrategy with an array of data",
+  fn() {
+    class Test extends Serializable {
+      @SerializeProperty({
+        toJSONStrategy: (v) => v.join("."),
+      })
+      public test_property = [0, 1];
+    }
+    assertEquals(new Test().toJSON(), `{"test_property":"0.1"}`);
   },
 });
 
@@ -248,7 +262,7 @@ test({
   name: "Errors on duplicate map keys",
   fn() {
     try {
-      class Test extends Serializable {
+      class _Test extends Serializable {
         @SerializeProperty("serialize_me")
         serializeMe = "nice";
         @SerializeProperty("serialize_me")
@@ -391,6 +405,7 @@ test({
     class Test2 extends Serializable {
       @SerializeProperty({
         serializedKey: "outer_property",
+        toJSONStrategy: (values: Test1[]) => values.map((v) => v.tsSerialize()),
       })
       nested: Test1[] = [new Test1()];
     }
@@ -398,6 +413,7 @@ test({
     class Test3 extends Serializable {
       @SerializeProperty({
         serializedKey: "outer_outer_property",
+        toJSONStrategy: (values: Test2[]) => values.map((v) => v.tsSerialize()),
       })
       nested2: Test2[] = [new Test2()];
     }
@@ -444,7 +460,7 @@ test({
   fn() {
     class Test1 extends Serializable {
       @SerializeProperty(
-        ({ serializedKey: (propertyName) => `_${String(propertyName)}` }),
+        { serializedKey: (propertyName) => `_${String(propertyName)}` },
       )
       serializeMe = "nice1";
     }
@@ -476,9 +492,8 @@ test({
   name: "should be able to serialize a serializable without any properties",
   fn() {
     class TestSerializable extends Serializable {}
-
-    assertEquals(new TestSerializable().fromJSON({}), {});
     const serializedClass = new TestSerializable().fromJSON({});
+    assert(serializedClass instanceof TestSerializable);
     assertEquals(Object.keys(serializedClass).length, 0);
   },
 });