+++ strategy/from_json/to_serializable.ts
@@ -2,15 +2,15 @@
 
 import { JSONValue, Serializable } from "../../serializable.ts";
 import { FromJSONStrategy } from "../compose_strategy.ts";
-import { getNewSerializable, SerializableConstructor } from "../utils.ts";
+import { getNewSerializable } from "../utils.ts";
 
 /** revive data using `fromJSON` on a subclass type */
-export function toSerializable<T>(
-  type: SerializableConstructor<T>,
+export function toSerializable(
+  type: unknown,
 ): FromJSONStrategy {
-  return function _toSerializable(
+  return (
     value: JSONValue,
-  ): Serializable | Serializable[] {
+  ): Serializable | Serializable[] => {
     if (Array.isArray(value)) {
       return value.map((item) => getNewSerializable(type).fromJSON(item));
     }